Changed the subject slightly.
Let me know if you have any better ideas rather than just making it smaller.

When it was bigger if the subject was too long it would move the reply link and move the menu button. This is the best solution I can think of without it looking totally retarded. I thought about putting it to the left of the menu and hide buttons but it looks strange and out of place there.